pyproject.toml
[project]
name = "ensured"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.0.5",
    "langchain-anthropic>=1.0.3",
    "langchain-google-genai>=3.0.3",
    "langchain-openai>=1.0.2",
    "langgraph>=1.0.3",
    "deepagents",
    "pytest-playwright>=0.7.1",
    "python-dotenv>=1.2.1",
    "pillow>=10.0.0",
    "moondream",
    "cua-computer>=0.4.12",
    "rich>=14.2.0",
    "requests>=2.32.5",
]
